// model/documentResponse.ts
/**
 * FlowAccount Open API
 * FlowAccount.com โปรแกรมบัญชีออนไลน์ใช้งานง่าย สำหรับธุรกิจที่พึ่งเริ่มต้น
 *
 * The version of the OpenAPI document: 2-oas3
 * Contact: developer@flowaccount.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class DocumentResponse {
    /**
    * action success
    */
    'status'?: boolean;
    /**
    * error message
    */
    'message'?: string;
    /**
    * error code
    */
    'code'?: number;
    'documentStructureType'?: string;

    static discriminator: string | undefined = "documentStructureType";

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "boolean"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "documentStructureType",
            "baseName": "documentStructureType",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DocumentResponse.attributeTypeMap;
    }
}