// model/expensenlineDocumentResponse.ts
/**
 * FlowAccount Open API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2-oas3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ExpenseInlineDocument } from './expenseInlineDocument';

export class ExpensenlineDocumentResponse {
    /**
    * action success
    */
    'status'?: boolean;
    /**
    * error message
    */
    'message'?: string;
    /**
    * error code
    */
    'code'?: number;
    'data'?: ExpenseInlineDocument;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "boolean"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "number"
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "ExpenseInlineDocument"
        }    ];

    static getAttributeTypeMap() {
        return ExpensenlineDocumentResponse.attributeTypeMap;
    }
}